invisibility support * remove support for RInvisible altogether, and break many tests * introduce visibility flag in RContext (true by default), query it in REngine, and introduce VisibilityController interface * turn all builtins into visibility controllers, adopt visibility controllership in builtins and several other nodes
NOTE: Node classes that control visibility MUST implement the VisibilityController interface and call controlVisibility() in EACH of their specializations, or, if the node is implemented in terms of execute() methods only, in each of those. There may be exceptions, such as in Uninitialized...Node classes, which replace themselves and forward execution to another node.